requirements.txt
beautifulsoup4>=4.7.*
requests2>=2.16.*
PyPDF2>=1.26.*
Flask>=1.0.2